+++ contracts/external/VotingPowerManagerERC20UniV3.sol
@@ -1,37 +1,37 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.9;
 
-import '@openzeppelin/contracts/utils/structs/EnumerableSet.sol';
-import '@openzeppelin/contracts/utils/structs/EnumerableMap.sol';
+import "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
+import "@openzeppelin/contracts/utils/structs/EnumerableMap.sol";
 import {LibUniswapV3Math} from "../libraries/helpers/LibUniswapV3Math.sol";
 import {IERC20} from "../libraries/openzeppelin/IERC20.sol";
 import {SafeERC20} from "../libraries/openzeppelin/SafeERC20.sol";
 import {IVotingPower} from "../interfaces/decisionSystem/IVotingPower.sol";
 
 struct Slot0 {
-    // the current price
-    uint160 sqrtPriceX96;
-    // the current tick
-    int24 tick;
-    // the most-recently updated index of the observations array
-    uint16 observationIndex;
-    // the current maximum number of observations that are being stored
-    uint16 observationCardinality;
-    // the next maximum number of observations to store, triggered in observations.write
-    uint16 observationCardinalityNext;
-    // the current protocol fee as a percentage of the swap fee taken on withdrawal
-    // represented as an integer denominator (1/x)%
-    uint8 feeProtocol;
-    // whether the pool is locked
-    bool unlocked;
+  // the current price
+  uint160 sqrtPriceX96;
+  // the current tick
+  int24 tick;
+  // the most-recently updated index of the observations array
+  uint16 observationIndex;
+  // the current maximum number of observations that are being stored
+  uint16 observationCardinality;
+  // the next maximum number of observations to store, triggered in observations.write
+  uint16 observationCardinalityNext;
+  // the current protocol fee as a percentage of the swap fee taken on withdrawal
+  // represented as an integer denominator (1/x)%
+  uint8 feeProtocol;
+  // whether the pool is locked
+  bool unlocked;
 }
 
 interface IUniV3Pool {
-  function slot0() external returns(Slot0 memory);
+  function slot0() external returns (Slot0 memory);
 }
 
 interface INFPositionManager {
-  function factory() external returns(address);
+  function factory() external returns (address);
 
   function ownerOf(uint256 tokenId) external view returns (address owner);
 
@@ -76,7 +76,7 @@ contract StakeContractERC20UniV3 {
   // staker => staked amount
   mapping(address => uint256) private _stakedERC20GovToken;
   // Mapping from staker address to their (enumerable) set of staked NFtokens
-  mapping (address => EnumerableSet.UintSet) private _stakerNFTPositions;
+  mapping(address => EnumerableSet.UintSet) private _stakerNFTPositions;
   // nftPositionTokenID => votingPower
   mapping(uint256 => uint256) private _amountOfVotingPowerForNFTPosition;
 
@@ -90,9 +90,12 @@ contract StakeContractERC20UniV3 {
     nfPositionManager = INFPositionManager(_nfPositionManager);
     uniV3Factory = nfPositionManager.factory();
     governanceToken = _governanceToken;
-    uint amountOfPairTokens = _legalPairTokens.length;
-    require(amountOfPairTokens > 0 && amountOfPairTokens < 10, "No pair token is set for pool or more than 9.");
-    for (uint i = 0; i < amountOfPairTokens; i++) {
+    uint256 amountOfPairTokens = _legalPairTokens.length;
+    require(
+      amountOfPairTokens > 0 && amountOfPairTokens < 10,
+      "No pair token is set for pool or more than 9."
+    );
+    for (uint256 i = 0; i < amountOfPairTokens; i++) {
       legalPairTokens.add(_legalPairTokens[i]);
     }
   }
@@ -106,7 +109,7 @@ contract StakeContractERC20UniV3 {
    * @notice Before call have to approve
    * @return Returns amount of voting power staker gets for staked tokens.
    */
-  function stakeGovToken(uint256 _amount) public returns(uint256) {
+  function stakeGovToken(uint256 _amount) public returns (uint256) {
     // receive tokens from holder to stake contract
     IERC20(governanceToken).safeTransferFrom(msg.sender, address(this), _amount); // double check
     // account how much holders tokens are staked
@@ -129,11 +132,8 @@ contract StakeContractERC20UniV3 {
   /**
    * @return Returns amount of voting power staker loses after unstaking tokens.
    */
-  function unstakeGovToken(uint256 _amount) public returns(uint256) {
-    require(
-      _stakedERC20GovToken[msg.sender] >= _amount,
-      "Trying to unstake more than have."
-    );
+  function unstakeGovToken(uint256 _amount) public returns (uint256) {
+    require(_stakedERC20GovToken[msg.sender] >= _amount, "Trying to unstake more than have.");
     // reduce token holdings
     _stakedERC20GovToken[msg.sender] -= _amount;
     // take back voting power
@@ -147,7 +147,7 @@ contract StakeContractERC20UniV3 {
    * @notice Before call have to approve (make operator = address(this))
    * @return Returns amount of voting power staker gets for staked position.
    */
-  function stakeUniV3NFTPosition(uint256 tokenId) public returns(uint256) {
+  function stakeUniV3NFTPosition(uint256 tokenId) public returns (uint256) {
     require(nfPositionManager.ownerOf(tokenId) == msg.sender, "Not an owner of NFT position.");
 
     (address operator, uint256 amountOfVotingPower) = _convertUNIV3PositionToVotingPower(tokenId);
@@ -169,8 +169,11 @@ contract StakeContractERC20UniV3 {
   /**
    * @return Returns amount of voting power staker loses after unstaking position.
    */
-  function unstakeUniV3NFTPosition(uint256 tokenId) public returns(uint256) {
-    require(_stakerNFTPositions[msg.sender].contains(tokenId), "tokenId was not staked by msg.sender");
+  function unstakeUniV3NFTPosition(uint256 tokenId) public returns (uint256) {
+    require(
+      _stakerNFTPositions[msg.sender].contains(tokenId),
+      "tokenId was not staked by msg.sender"
+    );
 
     // remove token from holdings
     _stakerNFTPositions[msg.sender].remove(tokenId);
@@ -189,21 +192,22 @@ contract StakeContractERC20UniV3 {
    * @notice Before call have to approve all (make operator = address(this))
    * @return Returns amount of voting power staker gets for staked positions.
    */
-  function stakeMultipleUniV3NFTPositions(uint256[] memory tokenIds) external returns(uint256) {
+  function stakeMultipleUniV3NFTPositions(uint256[] memory tokenIds) external returns (uint256) {
     require(tokenIds.length < 100);
     uint256 amountOfVotingPower;
-    for (uint i = 0; i < tokenIds.length; i++) {
+    for (uint256 i = 0; i < tokenIds.length; i++) {
       amountOfVotingPower += stakeUniV3NFTPosition(tokenIds[i]);
     }
     return amountOfVotingPower;
   }
+
   /**
    * @return Returns amount of voting power staker loses after unstaking positions.
    */
-  function unStakeMultipleUniV3NFTPositions(uint256[] memory tokenIds) external returns(uint256) {
+  function unStakeMultipleUniV3NFTPositions(uint256[] memory tokenIds) external returns (uint256) {
     require(tokenIds.length < 100);
     uint256 amountOfVotingPower;
-    for (uint i = 0; i < tokenIds.length; i++) {
+    for (uint256 i = 0; i < tokenIds.length; i++) {
       amountOfVotingPower += unstakeUniV3NFTPosition(tokenIds[i]);
     }
     return amountOfVotingPower;
@@ -222,28 +226,26 @@ contract StakeContractERC20UniV3 {
   }
 
   // Internal functions
-  function convertToPositionData(bytes memory data) internal pure returns(PositionData memory) {
+  function convertToPositionData(bytes memory data) internal pure returns (PositionData memory) {
     PositionReturnedData memory prd = abi.decode(data, (PositionReturnedData));
 
-    return PositionData({
-      operator: prd.operator,
-      token0: prd.token0,
-      token1: prd.token1,
-      tickLower: prd.tickLower,
-      tickUpper: prd.tickUpper,
-      liquidity: prd.liquidity,
-      fee: prd.fee
-    });
-
+    return
+      PositionData({
+        operator: prd.operator,
+        token0: prd.token0,
+        token1: prd.token1,
+        tickLower: prd.tickLower,
+        tickUpper: prd.tickUpper,
+        liquidity: prd.liquidity,
+        fee: prd.fee
+      });
   }
 
   // View functions
 
-  function getStakedBalanceOfGovernanceToken(address holder)
-    external
-    view
-    returns (uint256 balance)
-  {
+  function getStakedBalanceOfGovernanceToken(
+    address holder
+  ) external view returns (uint256 balance) {
     balance = _stakedERC20GovToken[holder];
   }
 
@@ -252,23 +254,31 @@ contract StakeContractERC20UniV3 {
     return _stakerNFTPositions[holder].length();
   }
 
-  function getNFTPositionIdOfHolderByIndex(address holder, uint256 index) public view returns (uint256) {
+  function getNFTPositionIdOfHolderByIndex(
+    address holder,
+    uint256 index
+  ) public view returns (uint256) {
     return _stakerNFTPositions[holder].at(index);
   }
 
-  function getAllNFTPositionIdsOfHolder(address holder) public view returns(uint256[] memory) {
+  function getAllNFTPositionIdsOfHolder(address holder) public view returns (uint256[] memory) {
     return _stakerNFTPositions[holder].values();
- }
+  }
 
-  function nftPositionIsStakedByHolder(address holder, uint256 tokenId) public view returns(bool) {
+  function nftPositionIsStakedByHolder(address holder, uint256 tokenId) public view returns (bool) {
     return _stakerNFTPositions[holder].contains(tokenId);
   }
 
-  function getAmountOfVotingPowerForNFTPosition(uint tokenId) external view returns(uint) {
+  function getAmountOfVotingPowerForNFTPosition(uint256 tokenId) external view returns (uint256) {
     return _amountOfVotingPowerForNFTPosition[tokenId];
   }
 
-  function onERC721Received(address operator, address from, uint256 tokenId, bytes calldata data) external pure returns (bytes4) {
+  function onERC721Received(
+    address operator,
+    address from,
+    uint256 tokenId,
+    bytes calldata data
+  ) external pure returns (bytes4) {
     return 0x150b7a02;
   }
 
@@ -279,26 +289,35 @@ contract StakeContractERC20UniV3 {
    * @return Returns position operator
    * @return Returns amount of voting power staker gets for staked position
    */
-  function isUniV3NFTValidView(uint256 tokenId) external returns(address, uint256) {
+  function isUniV3NFTValidView(uint256 tokenId) external returns (address, uint256) {
     return _convertUNIV3PositionToVotingPower(tokenId);
   }
 
   function _convertUNIV3PositionToVotingPower(
     uint256 tokenId
-  )
-    internal
-    returns(address operator, uint256 amountOfVotingPower)
-  {
-    (bool suc, bytes memory data) = address(nfPositionManager).call(abi.encodeWithSelector(0x99fbab88, tokenId));
+  ) internal returns (address operator, uint256 amountOfVotingPower) {
+    (bool suc, bytes memory data) = address(nfPositionManager).call(
+      abi.encodeWithSelector(0x99fbab88, tokenId)
+    );
     require(suc);
     PositionData memory positionData = convertToPositionData(data);
 
     operator = positionData.operator;
 
-    require(positionData.token0 == governanceToken || positionData.token1 == governanceToken, "No governance token in underlying assets.");
-    address pairToken = positionData.token0 == governanceToken ? positionData.token1 : positionData.token0;
+    require(
+      positionData.token0 == governanceToken || positionData.token1 == governanceToken,
+      "No governance token in underlying assets."
+    );
+    address pairToken = positionData.token0 == governanceToken
+      ? positionData.token1
+      : positionData.token0;
     require(legalPairTokens.contains(pairToken), "No legal pair token in underlying assets.");
-    address pool = LibUniswapV3Math.computePoolAddress(uniV3Factory, positionData.token0, positionData.token1, positionData.fee);
+    address pool = LibUniswapV3Math.computePoolAddress(
+      uniV3Factory,
+      positionData.token0,
+      positionData.token1,
+      positionData.fee
+    );
     Slot0 memory slot0 = IUniV3Pool(pool).slot0();
 
     uint160 sqrtRatioAX96 = LibUniswapV3Math.getSqrtRatioAtTick(positionData.tickLower);
@@ -308,7 +327,11 @@ contract StakeContractERC20UniV3 {
       if (positionData.token0 == governanceToken) {
         // here all position is HBT
         // amount0
-        amountOfVotingPower = LibUniswapV3Math.getAmount0ForLiquidity(sqrtRatioAX96, sqrtRatioBX96, positionData.liquidity);
+        amountOfVotingPower = LibUniswapV3Math.getAmount0ForLiquidity(
+          sqrtRatioAX96,
+          sqrtRatioBX96,
+          positionData.liquidity
+        );
       } else {
         // now we don't accept only WETH (other pair token) positions
         revert("Only pair token liquidity is not accepted yet.");
@@ -317,9 +340,17 @@ contract StakeContractERC20UniV3 {
       // here in range
       // we don't calcute the amount0 and amount1, instead we calculate amount if position would be out of range and contain only governanceToken
       if (positionData.token0 == governanceToken) {
-        amountOfVotingPower = LibUniswapV3Math.getAmount0ForLiquidity(sqrtRatioAX96, sqrtRatioBX96, positionData.liquidity);
+        amountOfVotingPower = LibUniswapV3Math.getAmount0ForLiquidity(
+          sqrtRatioAX96,
+          sqrtRatioBX96,
+          positionData.liquidity
+        );
       } else {
-        amountOfVotingPower = LibUniswapV3Math.getAmount1ForLiquidity(sqrtRatioAX96, sqrtRatioBX96, positionData.liquidity);
+        amountOfVotingPower = LibUniswapV3Math.getAmount1ForLiquidity(
+          sqrtRatioAX96,
+          sqrtRatioBX96,
+          positionData.liquidity
+        );
       }
     } else {
       if (positionData.token0 == governanceToken) {
@@ -328,9 +359,12 @@ contract StakeContractERC20UniV3 {
       } else {
         // here all position is HBT
         // amount1
-        amountOfVotingPower = LibUniswapV3Math.getAmount1ForLiquidity(sqrtRatioAX96, sqrtRatioBX96, positionData.liquidity);
+        amountOfVotingPower = LibUniswapV3Math.getAmount1ForLiquidity(
+          sqrtRatioAX96,
+          sqrtRatioBX96,
+          positionData.liquidity
+        );
       }
     }
   }
-
 }
