+++ contracts/external/token/InitialDistributor.sol
@@ -23,17 +23,14 @@ contract InitialDistributor {
     require(msg.sender == owner, "No rights to distribute");
     require(receivers.length == amounts.length, "Input length does not match");
     require(receivers.length <= 500, "No more than 500 per tx");
-    for (uint i = 0; i < receivers.length; i++) {
+    for (uint256 i = 0; i < receivers.length; i++) {
       distribute(receivers[i], amounts[i]);
     }
   }
 }
 
 interface IStakeERC20Contract {
-  function stakeGovInFavorOf(
-    address beneficiary,
-    uint256 _amount
-  ) external;
+  function stakeGovInFavorOf(address beneficiary, uint256 _amount) external;
 }
 
 contract InitialDistributorAbleToStake {
@@ -54,18 +51,22 @@ contract InitialDistributorAbleToStake {
     stakeERC20Contract = _stakeERC20Contract;
   }
 
-  function setTokenToDistribute(address _token) external returns(bool) {
+  function setTokenToDistribute(address _token) external returns (bool) {
     require(msg.sender == tokenSetter, "No rights to set contract");
     tokenToDistribute = _token;
     return true;
   }
 
-  function stakeTokensInFavorOfMultipleAddresses(address[] memory beneficiaries, uint[] memory amounts, uint totalAmount) external {
+  function stakeTokensInFavorOfMultipleAddresses(
+    address[] memory beneficiaries,
+    uint256[] memory amounts,
+    uint256 totalAmount
+  ) external {
     require(msg.sender == owner, "No rights to stake");
     require(beneficiaries.length == amounts.length, "input length does not match");
     IERC20(tokenToDistribute).approve(stakeERC20Contract, totalAmount);
-    uint amountStaked;
-    for (uint i = 0; i < beneficiaries.length; i++) {
+    uint256 amountStaked;
+    for (uint256 i = 0; i < beneficiaries.length; i++) {
       IStakeERC20Contract(stakeERC20Contract).stakeGovInFavorOf(beneficiaries[i], amounts[i]);
       amountStaked += amounts[i];
     }
@@ -80,7 +81,7 @@ contract InitialDistributorAbleToStake {
     require(msg.sender == owner, "No rights to distribute");
     require(receivers.length == amounts.length, "Input length does not match");
     require(receivers.length <= 500, "No more than 500 per tx");
-    for (uint i = 0; i < receivers.length; i++) {
+    for (uint256 i = 0; i < receivers.length; i++) {
       distribute(receivers[i], amounts[i]);
     }
   }
