+++ contracts/interfaces/IProposal.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.9;
 
 interface IProposal {
   struct Proposal {
-      bool proposalAccepted;
-      address destinationAddress;
-      uint value;
-      bytes callData;
-      bool proposalExecuted;
-      uint executionTimestamp;
-      //bytes ipfsHash;
-      //address proposer;
+    bool proposalAccepted;
+    address destinationAddress;
+    uint256 value;
+    bytes callData;
+    bool proposalExecuted;
+    uint256 executionTimestamp;
+    //bytes ipfsHash;
+    //address proposer;
   }
 }
